refactor: extract config normalization
